# delilah_Resto_Project/project/api.yaml
swagger: "2.0"
info:
  description: "This is the Delilah Restó Restaurant API. Where you can make your order online, and receive it in your home, easy and quick."
  version: "1.0.0"
  title: "Delilah Restó"
  termsOfService: "http://delilah/terms/"
  contact:
    email: "delilahresto@gmail.com.co"
  license:
    name: "Apache 2.0"
    url: "http://www.delilahresto.com.co"
host: "delilah.swagger.io"
basePath: "/v2"
...
schemes:
- "https"
- "http"
paths:
...
  /products:
    get:
      tags:
      - "products"
      summary: "Get a list of all the products availables"
      description: ""
      operationId: "getProduct"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "token"
        type: "string"
        description: "Token to authenticate."
        required: true
      responses:
        "200":
          description: "List of the all products availables"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/product"
        "403":
          description: "Session expired or User do not have permissions to access."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Expired"      
    post:
      tags:
      - "products"
      summary: "Add a new product to the list of products"
      description: ""
      operationId: "addPlate"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "product"
        description: "Product JSON object that needs to be added to the menu"
        required: true
        schema:
          $ref: "#/definitions/productPost"
      - in: "header"
        name: "token"
        type: "string"
        description: "Token to authenticate."
        required: true    
      responses:
        "201":
          description: "Product created"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/productResponse"
        "400":
          description: "Name, price and active are required."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Expired"      
        "409":
          description: "Product already exist"
          schema: 
            type: "string"
    put:
      tags:
      - "products"
      summary: "Update an existing product"
      description: ""
      operationId: "updatePlate"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "product"
        description: "Product JSON object that needs to be updated to the menu"
        required: true
        schema:
          $ref: "#/definitions/product"
      - in: "header"
        name: "token"
        type: "string"
        description: "Token to authenticate."
        required: true
      responses:
        "200":
          description: "Product has been updated."
          schema:
           type: "string"
        "400":
          description: "Product does not exist."
          schema:
           type: "string"
        "403":
          description: "Session expired or User do not have permissions to access."
          schema:
           type: "array"
           items:
             $ref: "#/definitions/Expired"
...
  /orders:
    get:
      tags:
      - "orders"
      summary: "Get all orders of the users"
      description: "Return a record of all orders of the user URL example: http://localhost:3000/orders?includeUser=true&includeOrderStatus=true&includePaymentMethod=true. Manager can get orders of all users."
      operationId: "getRecord"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "token"
        type: "string"
        description: "Token to authenticate."
        required: true
      responses:
        "200":
          description: "Successful operation"
          schema:
           type: "array"
           items:
             $ref: "#/definitions/order"
        "403":
          description: "Session expired or User do not have permissions to access."
          schema:
           type: "array"
           items:
             $ref: "#/definitions/Expired"
    post:
      tags:
      - "orders"
      summary: "Creates new order"
      description: "Creates a new order with (user id, order status id and payment method id)."
      operationId: "postOrder"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "order"
        description: "JSON object that needs to create a new order"
        required: true
        schema:
          $ref: "#/definitions/orderToPost"
      - in: "header"
        name: "token"
        type: "string"
        description: "Token to authenticate."
        required: true
      responses:
        "200":
          description: "Created"
          schema:
           type: "object"
           properties:
             order_id: 
               type: "integer"
        "400":
          description: "User_id, payment_method_id and order_status_id are required."
          schema:
           type: "object"
           properties:
             error:
              type: "string"
        "403":
          description: "Session expired or User do not have permissions to access."
          schema:
           type: "object"
           properties:
             error:
              type: "string"
    put:
      tags:
      - "orders"
      summary: "Manager update an order"
      description: "Update an specific order"
      operationId: "putOrder"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "order"
        description: "JSON object that needs to create a new order"
        required: true
        schema:
          type: "object"
          properties:
            id:
              type: "integer"
            order_status_id:
              type: "integer"
            payment_method_id:
              type: "integer"
      - in: "header"
        name: "token"
        type: "string"
        description: "Token to authenticate."
        required: true
      responses:
        "200":
          description: "Order has been updated"
          schema:
            type: "string"
        "400":
          description: "Id field is required"
          schema:
            type: "string"
        "403":
          description: "Session expired or User do not have permissions to access."
          schema:
            type: "object"
            properties:
             error:
              type: "string"    
        "404":
          description: "Order was not found"
          schema:
            type: "string"    
    delete:
      tags:
      - "orders"
      summary: "Delete an order by id (param)"
      description: "Cancel a specific order"
      operationId: "deleteOrder"
      produces:
      - "application/json"
      parameters:
      - in: "query"
        name: "id"
        type: "integer"
        description: "Id of order to delete."
        required: true
      - in: "header"
        name: "token"
        type: "string"
        description: "Token to authenticate."
        required: true
      responses:
        "204":
          description: "No content (deleted successful)."
        "403":
          description: "Session expired or User do not have permissions to access."
          schema:
            type: "object"
            properties:
             error:
              type: "string"  
        "404":
          description: "Order was not found."
          schema:
            type: "string"
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  Expired:
    type: "object"
    properties:
      error:
        type: "string"
        format: "int64"
    xml:
      name: "Expired"
...
  product:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string"
      price:
        type: "integer"
      active:
        type: "boolean"
    xml:
      name: "product"
  productPost:
    type: "object"
    properties:
      name:
        type: "string"
      price:
        type: "integer"
      active:
        type: "boolean"
    xml:
      name: "productPost"
  productResponse:
    type: "object"
    properties:
      product_id:
        type: "integer"
    xml:
      name: "productResponse"
  order:
    type: "object"
    properties:
      id:
        type: "integer"
      user_id:
        type: "integer"
      order_status_id:
        type: "integer"
      payment_method__id:
        type: "integer"  
      user:
        type: "object"
        properties:
          full_name:
            type: "string"
          address:
            type: "string"
          phone:
            type: "string"
          email:
            type: "string"
      oderStatus: 
       type: "object"
       properties:
         id:
           type: "integer"
         name:
           type: "string"
      paymentMethods:
       type: "object"
       properties:
         id:
           type: "integer"
         name:
           type: "string"
    xml:
      name: "order"
  orderToPost:
    type: "object"
    properties:
      user_id:
        type: "integer"
      order_status_id:
        type: "integer"
      payment_method_id:
        type: "integer"  
    xml:
      name: "orderToPost"    
...